调试工具

+++ include/kernel/task/sync.h
@@ -44,13 +44,13 @@ mutex_unlock(struct mutex_lock* lck);
 
 struct semaphore
 {
-  i8 value;
+  size_t value;
   struct spin_lock guard;
   struct list_head wait_queue;
 };
 
 void
-semaphore_init(struct semaphore* sem, i8 value);
+semaphore_init(struct semaphore* sem, size_t value);
 
 void
 semaphore_down(struct semaphore* sem);
